Allow attributes on type references in LimeIDL
Updated LimeIDL model classes to allow attributes on all elements,
instead of just named elements as before.

Updated LimeIDL grammar to allow annotations/attributes on type
references. This adds more flexibility for future language extensions.
In particular, it allows of implementing platform-specific semantics on
type references (see #268).

Extracted Antlr->LIME conversion functions for annotations/attributes
from AntlrLimeModelBuilder into a stand-alone class AntlrLimeConverter.

Updated AntlrTypeMapper to add attributes to type references it creates.
