+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeBasicTypeRef.kt
@@ -21,8 +21,9 @@ package com.here.gluecodium.model.lime
 
 class LimeBasicTypeRef(
     typeId: LimeBasicType.TypeId,
-    override val isNullable: Boolean = false
-) : LimeTypeRef() {
+    override val isNullable: Boolean = false,
+    attributes: LimeAttributes? = null
+) : LimeTypeRef(attributes) {
     override val elementFullName = typeId.name
     override val type = LimeBasicType(typeId)
