+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeLazyEnumeratorRef.kt
@@ -28,7 +28,7 @@ package com.here.gluecodium.model.lime
 class LimeLazyEnumeratorRef(
     referenceMap: Map<String, LimeElement>,
     override val elementFullName: String
-) : LimeEnumeratorRef {
+) : LimeEnumeratorRef() {
     override val enumerator by lazy {
         referenceMap[elementFullName] as? LimeEnumerator
             ?: throw LimeModelLoaderException("Enumerator $elementFullName was not found")
